Create parseTimeLimit util & Apply Detailpage #67
