+++ src/Pages/DetailPage/index.jsx
@@ -4,12 +4,14 @@ import PropTypes from 'prop-types';
 import { withRouter } from 'react-router-dom';
 import { connect } from 'react-redux';
 
-import { getPostByPostId } from 'Redux/post';
+import { getPostByPostId, getFundCount, getDueDate } from 'Redux/post';
 
 import NavBar from 'Components/Common/NavBar';
 
 import DueDate from 'Assets/images/dueDate-filled.svg';
 import HeartImg from 'Assets/images/fundHeart-bg-purple.png';
+
+import parseTimeLimit from 'Utils/parseTimeLimit';
 
 import profile1 from 'Dummies/dummyImage/1.jpg';
 import profile2 from 'Dummies/dummyImage/2.jpg';
@@ -33,6 +35,8 @@ class DetailPage extends Component {
     super(props);
     this.state = {
       post: {},
+      fundCount: 0,
+      dueDate: '',
     };
   }
 
@@ -60,9 +64,9 @@ class DetailPage extends Component {
 
     if (postId) {
       const detailedPost = await getPost(postId);
-      await new Promise((accept) =>
-        this.setState({ post: detailedPost }, accept),
-      );
+      const dueDate = parseTimeLimit(await getDueDate(postId));
+      const fundCount = await getFundCount(postId);
+      this.setState({ post: detailedPost, dueDate, fundCount });
     } else this.linkToNotFound();
 
     if (!(isActive === 'Ongoing' || isActive === 'End')) this.linkToNotFound();
@@ -71,6 +75,7 @@ class DetailPage extends Component {
   render() {
     const params = this.getParamsByQuery();
     const { isActive } = params;
+    const { dueDate, fundCount } = this.state;
     return (
       <div className="wrapper">
         <NavBar isActive="detail" />
@@ -95,13 +100,13 @@ class DetailPage extends Component {
                   <span>
                     <img src={DueDate} alt="dueDate" />
                   </span>
-                  <span>1:32:21</span>
+                  <span>{dueDate}</span>
                 </div>
                 <div>
                   <span>
                     <img src={HeartImg} alt="fund-heart" />
                   </span>
-                  <span>362</span>
+                  <span>{fundCount}</span>
                 </div>
               </div>
             )}
